realtime/requirements.txt
anthropic>=0.52.0
slack-bolt>=1.21.0
slack-sdk>=3.33.0
requests>=2.32.0
python-dotenv>=1.1.0
